src/api/pyproject.toml
[tool.poetry]
name = "slack-auth-api"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.73.0"
uvicorn = "^0.17.1"
autopep8 = "^1.6.0"
python-dotenv = "^0.19.2"
slack-sdk = "^3.13.0"
Jinja2 = "^3.0.3"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
